Split JSON utils
